working on anton integration
need to add the story to the right side of the page and the listener below the picture. to do this we need a generate story button (for now) then to call the socket listener to generate the story which should call the story generated listener to set the story in the function. which should then populate the page etc